pyproject.toml
[project]
name = "api-fetcher"
version = "0.1.0"
description = "This is poetry project for fetching an api"
authors = [
    {name = "Shvam Singh",email = "singhshivam583@yahoo.in"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
requests-mock = "^1.12.1"

